+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## unrelease
+
+### Changed
+- Metavariables can now contain digits or `_`. `$USERS_2` is now
+  a valid metavariable name. A metavariable must start with a letter
+  or `_` though.
+
+### Added
+- semgrep will return an error if your pattern contains an identifier 
+  starting with a dollar that is actually not considered a metavariable
+  (e.g., `$x`)
+
 ## [0.5.0](https://github.com/returntocorp/semgrep/releases/tag/v0.5.0) - 2020-04-28
 
 ### Changed